clean up mopro-ffi dep.
