+++ circom-prover/src/lib.rs
@@ -1,2 +1,25 @@
 pub mod prover;
 pub mod witness;
+
+pub use rust_witness;
+pub use rust_witness::transpile::transpile_wasm;
+pub use witnesscalc_adapter;
+
+#[derive(Debug, Clone, Default)]
+pub struct G1 {
+    pub x: String,
+    pub y: String,
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct G2 {
+    pub x: Vec<String>,
+    pub y: Vec<String>,
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct ProofCalldata {
+    pub a: G1,
+    pub b: G2,
+    pub c: G1,
+}
