+++ mopro-ffi/Cargo.toml
@@ -18,24 +18,19 @@ default = []
 ashlang = ["dep:ashlang"]
 halo2 = []
 circom = [
+    "circom-prover", 
     "rust-witness",
-    "ark-circom",
-    "ark-serialize",
-    "ark-ec",
-    "ark-crypto-primitives",
-    "ark-std",
     "ark-bn254",
-    "ark-groth16",
-    "ark-relations",
-    "ark-ff",
     "ark-bls12-381",
-    "num-traits",
-    "byteorder",
+    "ark-ff",
 ]
 
 [dependencies]
 uniffi = { version = "=0.28.0", features = ["cli", "build"] }
 serde = { version = "1.0", features = ["derive"] }
+anyhow = "1.0.86"
+bincode = "1.3.3"
+num-bigint = { version = "0.4.3", default-features = false, features = ["rand",] }
 
 # Error handling
 thiserror = "=2.0.3"
@@ -45,33 +40,14 @@ color-eyre = "=0.6.2"
 ashlang = { git = "https://github.com/chancehudson/ashlang.git", rev = "696960a0c15db47170fdd1ff058682023d904b1b", optional = true, default-features = false, features = ["spartan-prover", "serde"] }
 
 # circom deps
-ark-circom = { git = "https://github.com/zkmopro/circom-compat.git", version = "0.1.0", branch = "wasm-delete", optional = true }
-ark-serialize = { version = "=0.4.1", features = ["derive"], optional = true }
-num-bigint = { version = "0.4.3", default-features = false, features = [
-    "rand",
-] }
-circom-prover = {path = "../circom-prover"}
+rust-witness = { version = "0.1.1", optional = true }
+ark-ff = { version = "0.4.0", optional = true }
+circom-prover = {path = "../circom-prover", optional = true}
 
 # ZKP generation
-ark-ec = { version = "=0.4.1", default-features = false, features = [
-    "parallel",
-], optional = true }
-ark-crypto-primitives = { version = "=0.4.0", optional = true }
-ark-std = { version = "=0.4.0", default-features = false, features = [
-    "parallel",
-], optional = true }
 ark-bn254 = { version = "=0.4.0", optional = true }
-ark-groth16 = { version = "=0.4.0", default-features = false, features = [
-    "parallel",
-], optional = true }
-ark-relations = { version = "0.4", default-features = false, optional = true }
 uuid = { version = "1.9.1", features = ["v4"] }
-byteorder = { version = "1.0.0", optional = true }
-ark-ff = { version = "0.4.0", optional = true }
 ark-bls12-381 = { version = "0.4.0", optional = true }
-num-traits = { version = "0.2.0", optional = true }
-anyhow = "1.0.86"
-bincode = "1.3.3"
 
 [build-dependencies]
 rust-witness = { version = "0.1.1", optional = true }
